requirements.txt
grpcio==1.64.0
nltk==3.8.1
protobuf==5.27.0
psycopg2_binary==2.9.9
pyinflect==0.5.1
setuptools==58.0.4
setuptools==70.0.0
spacy==3.7.4
